postprocess: Display a progress bar for commits
This doesn't quite work as the count of files installed via RPM is
wildly off from the number we end up committing.  Also, hooking this
into xattr callback is a hack - we should improve ostree itself to do
this.

And if we were doing it right, we'd probably need to change ostree
to make two passes:
 - Once where we simply count files
 - Another where we commit