WIP: multiprocessing
I've been wanting to do this for a long time, but was too intimidated
at the thought of doing it in C.

One big thing to do here is to move the transaction heavy lifting
into a subprocess.  That would be a lot of work; just forking
off the dnf fetching would help solve
https://github.com/projectatomic/rpm-ostree/issues/897

Another case where I want do to this is to drop privileges for
the curl bits we do now for `rpm-ostree install http://`.

There's not a lot here yet, just me experimenting with Servo's
"ipc-channel".