Add `compose image`
This is also aliased as `dnf image build image-from-treefile`.  This new
flow avoids any need to know or understand ostree; one goes directly
from a treefile to an OCI image.

(Note the `from-treefile` suffixing leaves space for us to create
 a new non-"treefile" syntax)

Now internally, this is right now just a wrapper for the combination
of `rpm-ostree compose tree` which operates on a transient ostree
repository and `rpm-ostree compose container-encapsulate` which
turns that ostree commit into a reproducible OCI archive.

Crucially, we do maintain two important features here:

- change detection
- `automatic-version-prefix`

To implement this we fetch the prior container image metadata.
For the first build, it's required to pass `--initialize`.

I plan to change coreos-assembler to use this; it *almost* works
to just do `yum image build src/config/manifest.yaml` today, except
for the dynamic overlay bits.  We need to lower that into rpm-ostree.