container-encapsulate: make build_mapping_recurse significantly faster
While toying around with building my own custom FCOS builds, I noticed that
running `cosa build container` with a package set similar to Silverblue's
resulted in ~2hr builds, the vast majority of which was in the
"Building package mapping" task. After this change, the runtime on
my build shrank to ~15 mins.

`$ time cosa build container`
**Before**
```
real    10m47.769s
user    52m14.763s
sys     46m38.546s
```

**After**
```
real    15m37.333s
user    2m38.751s
sys     0m14.410s
```

The speedup is accomplished by avoiding the need to query the rpmdb for every file.
Instead the rpmdb is walked to build a cache of the files to providing packages,
so that when the ostree filesystem is walked later it can just check the cache.
The cache is structured similarly to rpm's internals, where paths are maintained
as separate basename and dirname entries. Additionally, like rpm, the paths
are considered equivalent if the dirnames resolve to the same path
(rpm uses `stat` to compare inodes, this implementation resolves the
symlinks). This results in output that is effectively equivalent to the
previous implementation while being substantially faster.

To minimize memory overhead maintaining the file mapping, a simple string cache is also added.

Closes: https://github.com/coreos/rpm-ostree/issues/4880